Add files via upload
How do I interpret user needs and implement them into a program? How does creating “user stories” help with this?
To interpret a user stories we have to understand their requirements, and preferences and make adjustment to the program. Creating user stories help by providing a clear view of what the user wants, and it helps making sure the final product meets the user expectations.

What does it mean to be a good team member in software development?

The traits that will contribute to being a good team member is communication and empathy it will help with the team success in finishing the product on time but also make the environment enjoyable to work in.